+++ src/components/feedback-form/components/finished/finished.module.css
@@ -1,0 +1,29 @@
+.finished {
+	composes: hds-typography-body-200 from global;
+	composes: animate;
+	width: 100%;
+	display: flex;
+	align-items: center;
+	gap: 12px;
+	font-weight: var(--token-typography-font-weight-semibold);
+	padding: 16px;
+	color: var(--token-color-foreground-success-on-surface);
+	background-color: var(--token-color-surface-success);
+	border: 1px solid var(--token-color-border-success);
+	border-radius: 6px;
+	animation: nextFrame 0.4s;
+	animation-fill-mode: both;
+	animation-timing-function: ease-in-out;
+}
+
+@keyframes nextFrame {
+	0% {
+		opacity: 0;
+		transform: translateY(50px);
+	}
+
+	100% {
+		opacity: 1;
+		transform: translateY(0);
+	}
+}
