+++ src/components/feedback-form/components/index.ts
@@ -1,0 +1,2 @@
+export * from './finished'
+export * from './question'
