+++ src/pages/api/tutorial-feedback.ts
@@ -8,7 +8,7 @@ import {
 import Bowser from 'bowser'
 import Bugsnag from '@bugsnag/js'
 
-const FEEDBACK_SHEET_ID = process.env.FEEDBACK_SHEET_ID
+const FEEDBACK_SHEET_ID = process.env.TUTORIAL_FEEDBACK_SHEET_ID
 const FEEDBACK_SERVICE_EMAIL = process.env.FEEDBACK_SERVICE_EMAIL
 const FEEDBACK_PRIVATE_KEY = process.env.FEEDBACK_PRIVATE_KEY
 const HASHI_ENV = process.env.HASHI_ENV
